wip compilation with jdk17

+++ node-api-tests/src/test/kotlin/net/corda/nodeapitests/internal/serialization/kryo/KryoAttachmentTest.kt
@@ -49,7 +49,7 @@ class KryoAttachmentTest(private val compression: CordaSerializationEncoding?) {
 
     @Test(timeout=300_000)
     fun `HashCheckingStream (de)serialize`() {
-        val rubbish = ByteArray(12345) { (it * it * 0.12345).toByte() }
+        val rubbish = ByteArray(12345) { (it * it * 0.12345).toInt().toByte() }
         val readRubbishStream: InputStream = NodeAttachmentService.HashCheckingStream(
                 SecureHash.sha256(rubbish),
                 rubbish.size,
